+++ Cargo.toml
@@ -9,23 +9,17 @@ members = [
     "crates/buoy_physics",
     "crates/buoy_atmo",
     "crates/buoy_aero",
-    "crates/buoy_sim",
     
     # Application
-    "crates/buoy_app",
+    "crates/buoy_sim",
 ]
 
 default-members = [
     "crates/buoy_sim",
 ]
 
-# Default binary to run when using `cargo run` from workspace root
 [workspace.metadata.bin]
 default = "buoy_sim"
-
-# Alternatively, you can set the default member to run
-# This makes `cargo run` from root run the buoy-server binary
-default-run = "buoy_sim"
 
 [workspace.package]
 version = "0.1.0"
@@ -67,6 +61,7 @@ tracing = { version = "0.1.41", features = [
 ] }
 nalgebra = "0.33.2"
 serde = "1.0.219"
+serde_json = "1.0.107"
 
 # TIP FROM TheBevyFlock/bevy_new_2d:
 # Your web builds will start failing if you add a dependency that pulls in `getrandom` v0.3+.
