+++ crates/buoy_common/src/constants.rs
