+++ crates/buoy_common/src/geometry.rs
