--- crates/buoy_core/Cargo.toml
@@ -1,38 +1,0 @@
-[package]
-name = "buoy_core"
-version = "0.1.0"
-edition = "2024"
-description = "Legacy core physics and simulation code"
-
-[lib]
-path = "src/lib.rs"
-
-[dependencies]
-bevy = { workspace = true }
-nalgebra = { workspace = true }
-serde = { workspace = true }
-avian3d = { workspace = true }
-uom = { workspace = true }
-log = { workspace = true }
-tracing = { workspace = true }
-
-[features]
-default = [
-    "dev",
-]
-dev = [
-    "bevy/bevy_debug_stepping",
-]
-
-[lints.rust]
-# Mark `bevy_lint` as a valid `cfg`, as it is set when the Bevy linter runs.
-unexpected_cfgs = { level = "warn", check-cfg = ["cfg(bevy_lint)"] }
-
-[lints.clippy]
-# Bevy supplies arguments to systems via dependency injection, so it's natural for systems to
-# request more than 7 arguments, which would undesirably trigger this lint.
-too_many_arguments = "allow"
-# Queries may access many components, which would undesirably trigger this lint.
-type_complexity = "allow"
-# Make sure macros use their standard braces, such as `[]` for `bevy_ecs::children!`.
-nonstandard_macro_braces = "warn"
