+++ crates/buoy_physics/Cargo.toml
@@ -9,10 +9,22 @@ readme = "README.md"
 path = "src/lib.rs"
 
 [dependencies]
+bevy = { workspace = true }
 uom = { workspace = true }
 thiserror = { workspace = true }
 nalgebra = { workspace = true }
 serde = { workspace = true }
+avian3d = { workspace = true }
+log = { workspace = true }
+tracing = { workspace = true }
+
+[features]
+default = [
+    "dev",
+]
+dev = [
+    "bevy/bevy_debug_stepping",
+]
 
 [dev-dependencies]
 proptest = { workspace = true }
